requirements.txt
python-fasthtml
monsterui
openai
python-dotenv